modify login component

+++ frontend/src/app/service/token-storage.service.ts
@@ -1,4 +1,5 @@
 import {Injectable} from '@angular/core';
+import {User} from "../models/User";
 
 const TOKEN_KEY = 'auth-token';
 const USER_KEY = 'auth-user';
@@ -8,29 +9,43 @@ const USER_KEY = 'auth-user';
 })
 export class TokenStorageService {
 
-  constructor() {
-  }
+  constructor() { }
 
   public saveToken(token: string): void {
-    window.sessionStorage.removeItem(TOKEN_KEY);
-    window.sessionStorage.setItem(TOKEN_KEY, token);
+    if (typeof sessionStorage !== 'undefined') {
+      window.sessionStorage.removeItem(TOKEN_KEY);
+      window.sessionStorage.setItem(TOKEN_KEY, token);
+    }
   }
 
-  public getToken(): string {
-    return sessionStorage.getItem(TOKEN_KEY)!;
+  public getToken(): string | null {
+    if (typeof sessionStorage !== 'undefined') {
+      return sessionStorage.getItem(TOKEN_KEY);
+    }
+    return null;
   }
 
-  public saveUser(user:any): void {
-    window.sessionStorage.removeItem(USER_KEY);
-    window.sessionStorage.setItem(USER_KEY, JSON.stringify(user));
+  public saveUser(user: User): void {
+    if (typeof sessionStorage !== 'undefined') {
+      window.sessionStorage.removeItem(USER_KEY);
+      window.sessionStorage.setItem(USER_KEY, JSON.stringify(user));
+    }
   }
 
-  public getUser(): any {
-    return JSON.parse(sessionStorage.getItem(USER_KEY)!);
+  public getUser(): User | null {
+    if (typeof sessionStorage !== 'undefined') {
+      const userJson = sessionStorage.getItem(USER_KEY);
+      if (userJson !== null) {
+        return JSON.parse(userJson);
+      }
+    }
+    return null;
   }
 
-  logOut(): void {
-    window.sessionStorage.clear();
-    window.location.reload();
+  public logOut(): void {
+    if (typeof sessionStorage !== 'undefined') {
+      window.sessionStorage.clear();
+      window.location.reload();
+    }
   }
 }
